Refactor keyboard sensor tests.
