+++ tests/src/keyboard-sensor/events/index.ts
@@ -1,0 +1,13 @@
+import { eventCancel } from './cancel.js';
+import { eventEnd } from './end.js';
+import { eventMove } from './move.js';
+import { eventStart } from './start.js';
+
+export function events() {
+  describe('events', () => {
+    eventCancel();
+    eventEnd();
+    eventMove();
+    eventStart();
+  });
+}
