+++ tests/src/keyboard-sensor/methods/on.ts
@@ -1,0 +1,83 @@
+import { assert } from 'chai';
+import { createTestElement } from '../../utils/create-test-element.js';
+import { focusElement } from '../../utils/focus-element.js';
+import { KeyboardSensor } from '../../../../src/index.js';
+
+export function methodOn() {
+  describe('on', () => {
+    it('should return a unique symbol by default', () => {
+      const el = createTestElement();
+      const s = new KeyboardSensor(el);
+      const idA = s.on('start', () => {});
+      const idB = s.on('start', () => {});
+      assert.equal(typeof idA, 'symbol');
+      assert.notEqual(idA, idB);
+      el.remove();
+      s.destroy();
+    });
+
+    it('should allow duplicate event listeners', () => {
+      const el = createTestElement();
+      const s = new KeyboardSensor(el);
+      let counter = 0;
+      const listener = () => {
+        ++counter;
+      };
+
+      s.on('start', listener);
+      s.on('start', listener);
+
+      focusElement(el);
+      document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Enter' }));
+
+      assert.equal(counter, 2);
+
+      el.remove();
+      s.destroy();
+    });
+
+    it('should remove the existing listener and add the new one if the same id is used', () => {
+      const el = createTestElement();
+      const s = new KeyboardSensor(el);
+      let msg = '';
+
+      s.on('start', () => void (msg += 'a'), 1);
+      s.on('start', () => void (msg += 'b'), 2);
+      s.on('start', () => void (msg += 'c'), 1);
+
+      focusElement(el);
+      document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Enter' }));
+
+      assert.equal(msg, 'bc');
+
+      el.remove();
+      s.destroy();
+    });
+
+    it('should allow defining a custom id (string/symbol/number) for the event listener via third argument', () => {
+      const el = createTestElement();
+      const s = new KeyboardSensor(el);
+
+      const idA = Symbol();
+      assert.equal(
+        s.on('start', () => {}, idA),
+        idA,
+      );
+
+      const idB = 1;
+      assert.equal(
+        s.on('start', () => {}, idB),
+        idB,
+      );
+
+      const idC = 'foo';
+      assert.equal(
+        s.on('start', () => {}, idC),
+        idC,
+      );
+
+      el.remove();
+      s.destroy();
+    });
+  });
+}
