+++ tests/src/keyboard-sensor/properties/index.ts
@@ -1,0 +1,9 @@
+import { propDrag } from './drag.js';
+import { propIsDestroyed } from './is-destroyed.js';
+
+export function properties() {
+  describe('properties', () => {
+    propDrag();
+    propIsDestroyed();
+  });
+}
